chore: move utils/general.ts -> lib/general.ts
