+++ pages/api/roadmap.ts
@@ -5,7 +5,7 @@ import _, { result } from 'lodash';
 import { getIssue } from '../../lib/backend/issue';
 import { getChildren, getConfig } from '../../lib/parser';
 import { IssueData, ParserGetChildrenResponse, RoadmapApiResponse, RoadmapApiResponseFailure, RoadmapApiResponseSuccess } from '../../lib/types';
-import { paramsFromUrl } from '../../utils/general';
+import { paramsFromUrl } from '../../lib/general';
 import { errorManager } from '../../lib/backend/errorManager';
 
 const resolveChildren = (children: any[]): Promise<IssueData[]> => {
